Joystick Fix and Player Animations

+++ Artifice/app/src/main/java/com/underwaterotter/artifice/entities/main/CharController.java
@@ -8,7 +8,7 @@ public class CharController {
         GameScene.scene.player.sprite.angle = angle;
     }
 
-    public static void setSpeed(int speed){
+    public static void setSpeed(float speed){
         GameScene.scene.player.speed = speed;
     }
